+++ cmd/scheduler/app/server.go
@@ -90,6 +90,7 @@ func BuildSchedulerParams(opt *options.ServerOption) *conf.SchedulerParams {
 		GlobalDefaultStalenessGracePeriod: opt.GlobalDefaultStalenessGracePeriod,
 		SchedulePeriod:                    opt.SchedulePeriod,
 		DetailedFitErrors:                 opt.DetailedFitErrors,
+		UpdatePodEvictionCondition:        opt.UpdatePodEvictionCondition,
 	}
 }