+++ pkg/scheduler/cache/cache.go
@@ -91,6 +91,7 @@ type SchedulerCacheParams struct {
 	NodeLevelScheduler          bool
 	AllowConsolidatingReclaim   bool
 	NumOfStatusRecordingWorkers int
+	UpdatePodEvictionCondition  bool
 }
 
 type SchedulerCache struct {
@@ -143,7 +144,7 @@ func newSchedulerCache(schedulerCacheParams *SchedulerCacheParams) *SchedulerCac
 	broadcaster.StartRecordingToSink(&corev1.EventSinkImpl{Interface: sc.kubeClient.CoreV1().Events("")})
 	recorder := broadcaster.NewRecorder(kubeaischedulerschema.Scheme, v1.EventSource{Component: schedulerName})
 
-	sc.Evictor = evictor.New(sc.kubeClient)
+	sc.Evictor = evictor.New(sc.kubeClient, schedulerCacheParams.UpdatePodEvictionCondition)
 
 	sc.StatusUpdater = status_updater.New(
 		sc.kubeClient, sc.kubeAiSchedulerClient, recorder, schedulerCacheParams.NumOfStatusRecordingWorkers,
@@ -243,7 +244,7 @@ func (sc *SchedulerCache) evict(evictedPod *v1.Pod, evictedPodGroup *enginev2alp
 
 		log.InfraLogger.V(6).Infof("Evicting pod %v/%v, reason: %v, message: %v",
 			evictedPod.Namespace, evictedPod.Name, status.Preempted, message)
-		err := sc.Evictor.Evict(evictedPod)
+		err := sc.Evictor.Evict(evictedPod, message)
 		if err != nil {
 			log.InfraLogger.Errorf("Failed to evict pod: %v/%v, error: %v", evictedPod.Namespace, evictedPod.Name, err)
 		}